Move BasicApiTransformer to a separate file. test=develop
